+++ common/utils.h
@@ -14,7 +14,7 @@
 #define MIN(a,b) (((a)<(b))?(a):(b))
 
 float log2f( float n );
-float log2( float n );
+double log2( double n );
 
 uint16_t satAddU16U16(uint16_t a, uint16_t b);
 uint16_t satAddU16S32(uint16_t a, int32_t b);
